backend/tests/test_auth.py
import pytest
from httpx import AsyncClient
from unittest.mock import patch, MagicMock
from app.main import app # Assuming app.main refers to your FastAPI app instance
from app.core.config import settings

@pytest.fixture(scope="module")
def anyio_backend():
    return "asyncio"

@pytest.fixture(scope="module")
async def client():
    async with AsyncClient(app=app, base_url="http://test") as client:
        yield client

@patch.object(settings, "GOOGLE_CLIENT_ID", "test_client_id")
@patch.object(settings, "SUPABASE_REDIRECT_URI", "http://localhost:8000/api/v1/auth/google/callback")
async def test_google_oauth_initiate_redirects_correctly(client: AsyncClient):
    response = await client.get("/api/v1/auth/google")
    assert response.status_code == 307 # Temporary Redirect

    redirect_url = response.headers["location"]
    assert "https://accounts.google.com/o/oauth2/v2/auth" in redirect_url
    assert "client_id=test_client_id" in redirect_url
    assert "redirect_uri=http%3A%2F%2Flocalhost%3A8000%2Fapi%2Fv1%2Fauth%2Fgoogle%2Fcallback" in redirect_url
    assert "response_type=code" in redirect_url
    assert "scope=openid%20email%20profile" in redirect_url
    assert "access_type=offline" in redirect_url
    assert "state=some_random_secure_string" in redirect_url

async def test_google_oauth_callback_placeholder(client: AsyncClient):
    # This test is a placeholder to demonstrate the structure for the callback.
    # A full implementation would involve extensive mocking of external services (Google and Supabase).

    # Mock Google token exchange
    with patch("httpx.AsyncClient.post") as mock_google_post:
        mock_google_post.return_value = MagicMock(
            status_code=200,
            json=lambda: {"access_token": "google_test_access_token", "expires_in": 3600},
            raise_for_status=lambda: None
        )

        # Mock Supabase authentication
        with patch("supabase.Client.auth.sign_in_with_oauth") as mock_supabase_oauth:
            mock_supabase_oauth.return_value = MagicMock(
                # Supabase's sign_in_with_oauth returns a response that might contain session info or redirect URL
                # For this test, we'll mock a successful session creation
                session=MagicMock(access_token="supabase_test_jwt"),
                user=MagicMock(id="test_user_id")
            )
            
            response = await client.get("/api/v1/auth/google/callback", params={"code": "test_code", "state": "some_random_secure_string"})
            
            # Assert redirection to frontend base URL
            assert response.status_code == 307
            assert response.headers["location"] == settings.BASE_URL

            # Verify mocks were called
            mock_google_post.assert_called_once()
            mock_supabase_oauth.assert_called_once()
            # Further assertions would verify the arguments passed to mocks and the content of the session.
